check for non empty dir

+++ operator/src/main/resources/scripts/startServer.sh
@@ -222,8 +222,10 @@ function prepareMIIServer() {
         cd ${DOMAIN_HOME}/lib || return 1
         ${JAVA_HOME}/bin/jar xf ${IMG_ARCHIVES_ROOTDIR}/${file} wlsdeploy/domainLibraries/
         # Flatten the jars to the domain lib root
-        mv wlsdeploy/domainLibraries/* .
-        rm -fr wlsdeploy/
+        if [ "$(ls -A wlsdeploy/domainLibraries/)" ]; then
+          mv wlsdeploy/domainLibraries/* .
+          rm -fr wlsdeploy/
+        fi
         if [ $? -ne 0 ] ; then
           trace SEVERE  "Domain Source Type is FromModel, error in extracting domain libs ${IMG_ARCHIVES_ROOTDIR}/${file}"
           return 1
